pyproject.toml
[project]
name = "hallucination-inspector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.4",
    "langchain-openai>=0.3.27",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "uqlm>=0.1.7",
]
